# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "AnimeThemes.moe Api Documentation",
        "description": "AnimeThemes.moe Laravel RESTful API Documentation",
        "version": "1.0.0"
    },
    "paths": {
...
        "/entry/": {
            "get": {
                "tags": [
                    "Entry"
                ],
                "summary": "Get paginated listing of Entries",
                "description": "Returns listing of Entries",
                "operationId": "getEntries",
                "parameters": [
                    {
                        "name": "q",
                        "in": "query",
                        "description": "The search query. Mapping is to [entry.theme.anime.name|entry.theme.anime.synonyms.text + entry.theme.slug + entry.version] or entry.theme.song.title.",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "bakemonogatari ED"
                    },
                    {
                        "name": "include",
                        "in": "query",
                        "description": "Comma-separated list of included related resources. Allowed list is anime, themes & videos.",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "anime,videos"
                    },
                    {
                        "name": "version",
                        "in": "query",
                        "description": "Filter entries by version.",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 2
                    },
                    {
                        "name": "nsfw",
                        "in": "query",
                        "description": "Filter entries by NSFW. Case-insensitive options for true are 1, true, on & yes. Case-insensitive options for false are 0, false, off & no.",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        },
                        "example": "false"
                    },
                    {
                        "name": "spoiler",
                        "in": "query",
                        "description": "Filter entries by Spoiler. Case-insensitive options for true are 1, true, on & yes. Case-insensitive options for false are 0, false, off & no.",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        },
                        "example": "false"
                    },
                    {
                        "name": "order",
                        "in": "query",
                        "description": "Order entries by field. Case-insensitive options are entry_id, created_at, updated_at, version, episodes, nsfw, spoiler, notes & theme_id.",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "updated_at"
                    },
                    {
                        "name": "direction",
                        "in": "query",
                        "description": "Direction of entry ordering. Case-insensitive options are asc & desc.",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "desc"
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "The number of resources to return per page. Acceptable range is [1-100]. Default value is 100.",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 50
                    },
                    {
                        "name": "fields",
                        "in": "query",
                        "description": "The comma-separated list of fields by resource type",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "fields[entry]=version"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "entries": {
                                            "description": "Display a listing of the resource.",
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/EntryResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/entry/{id}": {
            "get": {
                "tags": [
                    "Entry"
                ],
                "summary": "Get properties of Entry",
                "description": "Returns properties of Entry",
                "operationId": "getEntry",
                "parameters": [
                    {
                        "name": "include",
                        "in": "query",
                        "description": "Comma-separated list of included related resources. Allowed list is anime, themes & videos.",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "anime,videos"
                    },
                    {
                        "name": "fields",
                        "in": "query",
                        "description": "The comma-separated list of fields by resource type",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "fields[entry]=version"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/EntryResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Resource Not Found!"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "EntryResource": {
                "title": "Entry",
                "description": "Entry Resource",
                "properties": {
                    "id": {
                        "description": "Primary Key",
                        "type": "integer",
                        "example": 3518
                    },
                    "version": {
                        "description": "The Version number of the Theme",
                        "type": "integer",
                        "example": ""
                    },
                    "episodes": {
                        "description": "The range(s) of episodes that the theme entry is used",
                        "type": "string",
                        "example": "1-2, 12"
                    },
                    "nsfw": {
                        "description": "Does the entry include Not Safe For Work content?",
                        "type": "bool",
                        "example": "false"
                    },
                    "spoiler": {
                        "description": "Does the entry include content that spoils the show?",
                        "type": "bool",
                        "example": "false"
                    },
                    "notes": {
                        "description": "Any additional information not included in other fields that may be useful",
                        "type": "string",
                        "example": ""
                    },
                    "created_at": {
                        "description": "The Resource Creation Timestamp",
                        "type": "string",
                        "example": "2020-08-15T05:43:02.000000Z"
                    },
                    "updated_at": {
                        "description": "The Resource Last Updated Timestamp",
                        "type": "string",
                        "example": "2020-08-15T05:43:02.000000Z"
                    },
                    "anime": {
                        "properties": {
                            "id": {
                                "description": "Primary Key",
                                "type": "integer",
                                "example": 197
                            },
                            "name": {
                                "description": "The Primary Title of the Anime",
                                "type": "string",
                                "example": "Bakemonogatari"
                            },
                            "alias": {
                                "description": "URL Slug & Model Route Key",
                                "type": "string",
                                "example": "bakemonogatari"
                            },
                            "year": {
                                "description": "The Year in which the Anime Premiered",
                                "type": "integer",
                                "example": 2009
                            },
                            "season": {
                                "description": "The Season in which the Anime Premiered",
                                "type": "string",
                                "enum": [
                                    "Winter",
                                    "Spring",
                                    "Summer",
                                    "Fall"
                                ],
                                "example": "Summer"
                            },
                            "created_at": {
                                "description": "The Resource Creation Timestamp",
                                "type": "string",
                                "example": "2020-08-15T05:30:43.000000Z"
                            },
                            "updated_at": {
                                "description": "The Resource Last Updated Timestamp",
                                "type": "string",
                                "example": "2020-08-15T05:37:25.000000Z"
                            }
                        },
                        "type": "object"
                    },
                    "theme": {
                        "properties": {
                            "id": {
                                "description": "Primary Key",
                                "type": "integer",
                                "example": 3102
                            },
                            "type": {
                                "description": "Is this an OP or an ED?",
                                "type": "string",
                                "enum": [
                                    "OP",
                                    "ED"
                                ],
                                "example": "OP"
                            },
                            "sequence": {
                                "description": "Numeric ordering of theme",
                                "type": "integer",
                                "example": "1"
                            },
                            "group": {
                                "description": "For separating sequences belonging to dubs, rebroadcasts, remasters, etc",
                                "type": "string",
                                "example": ""
                            },
                            "slug": {
                                "description": "URL Slug",
                                "type": "bool",
                                "example": "OP1"
                            },
                            "created_at": {
                                "description": "The Resource Creation Timestamp",
                                "type": "string",
                                "example": "2020-08-15T05:43:02.000000Z"
                            },
                            "updated_at": {
                                "description": "The Resource Last Updated Timestamp",
                                "type": "string",
                                "example": "2020-08-15T05:43:02.000000Z"
                            }
                        },
                        "type": "object"
                    },
                    "videos": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "id": {
                                    "description": "Primary Key",
                                    "type": "integer",
                                    "example": 2615
                                },
                                "basename": {
                                    "description": "The basename of the Video",
                                    "type": "string",
                                    "example": "Bakemonogatari-OP1.webm"
                                },
                                "filename": {
                                    "description": "The filename of the Video",
                                    "type": "string",
                                    "example": "Bakemonogatari-OP1"
                                },
                                "path": {
                                    "description": "The path of the Video in storage",
                                    "type": "string",
                                    "example": "2009/Summer/Bakemonogatari-OP1.webm"
                                },
                                "resolution": {
                                    "description": "Frame height of the video",
                                    "type": "integer",
                                    "example": 1080
                                },
                                "nc": {
                                    "description": "Is the video creditless?",
                                    "type": "bool",
                                    "example": "true"
                                },
                                "subbed": {
                                    "description": "Does the video include subtitles of dialogue?",
                                    "type": "bool",
                                    "example": "false"
                                },
                                "lyrics": {
                                    "description": "Does the video include subtitles for song lyrics?",
                                    "type": "bool",
                                    "example": "false"
                                },
                                "uncen": {
                                    "description": "Is the video an uncensored version of a censored sequence?",
                                    "type": "bool",
                                    "example": "false"
                                },
                                "source": {
                                    "description": "Where did this video come from?",
                                    "type": "string",
                                    "enum": [
                                        "WEB",
                                        "RAW",
                                        "BD",
                                        "DVD",
                                        "VHS"
                                    ],
                                    "example": "BD"
                                },
                                "overlap": {
                                    "description": "The degree to which the sequence and episode content overlap",
                                    "type": "string",
                                    "enum": [
                                        "None",
                                        "Transition",
                                        "Over"
                                    ],
                                    "example": "None"
                                },
                                "created_at": {
                                    "description": "The Resource Creation Timestamp",
                                    "type": "string",
                                    "example": "2020-08-15T05:30:43.000000Z"
                                },
                                "updated_at": {
                                    "description": "The Resource Last Updated Timestamp",
                                    "type": "string",
                                    "example": "2020-08-15T05:30:43.000000Z"
                                },
                                "link": {
                                    "description": "The link to the video stream",
                                    "type": "string",
                                    "example": "https://animethemes.moe/video/Bakemonogatari-OP1.webm"
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
...
        }
    },
...
    "security": [
        []
    ]
}